[RTT] -Fix: accept that gcode won't export some of the test cases (no-copper)
git-svn-id: svn://repo.hu/pcb-rnd/trunk@27034 1d61de99-560f-4dba-8957-1e505d42358b

+++ tests/RTT/Export.sh
@@ -117,7 +117,7 @@ move_out()
 			' $raw_out
 			;;
 		gcode)
-			for n in $raw_out.*.cnc
+			for n in `ls $raw_out.*.cnc 2>/dev/null`
 			do
 			awk '
 				/^[(] / && (NR < 5) { next } # skip date
@@ -152,8 +152,12 @@ move_out()
 			;;
 		gcode)
 			mkdir -p $final_out
-			mv $raw_out.*.cnc $final_out
-			rm -f ${raw_out%%.gcode}.*.png
+			files=`ls $raw_out.*.cnc 2>/dev/null`
+			if test ! -z "$files"
+			then
+				mv $files $final_out
+				rm -f ${raw_out%%.gcode}.*.png
+			fi
 			;;
 		*)
 			# common, single file output
